tinycms analytics: use effect to request data in new components
