+++ components/tinycms/analytics/Report.js
@@ -3,6 +3,7 @@ import { addDays } from 'date-fns';
 import { getMetricsData } from '../../../lib/analytics';
 import { formatDate } from '../../../lib/utils';
 import NewsletterSignupFormData from './NewsletterSignupFormData';
+import ReadingDepthData from './ReadingDepthData';
 import ReadingFrequencyData from './ReadingFrequencyData';
 
 const Report = (props) => {
@@ -21,9 +22,6 @@ const Report = (props) => {
   const [pageViewReportData, setPageViewReportData] = useState(INITIAL_STATE);
   const [geoReportData, setGeoReportData] = useState(INITIAL_STATE);
   const [referralData, setReferralData] = useState(INITIAL_STATE);
-  const [readingDepthData, setReadingDepthData] = useState({});
-  const [sortedReadingDepth, setSortedReadingDepth] = useState([]);
-  const [readingDepthTableRows, setReadingDepthTableRows] = useState([]);
   const [startDate, setStartDate] = useState(addDays(new Date(), -30));
   const [endDate, setEndDate] = useState(new Date());
   const [average, setAverage] = useState(0);
@@ -121,69 +119,6 @@ const Report = (props) => {
     });
   };
 
-  let readingDepthRows = [];
-  const displayReadingDepthResults = (response, initialData, setData) => {
-    const queryResult = response.result.reports[0].data.rows;
-    // console.log('reading depth: ', queryResult);
-
-    let collectedData = {};
-    queryResult.forEach((row) => {
-      let articlePath = row.dimensions[3];
-
-      if (articlePath !== '/') {
-        let percentage = row.dimensions[1];
-        let count = row.metrics[0].values[0];
-        if (collectedData[articlePath]) {
-          collectedData[articlePath][percentage] = count;
-        } else {
-          collectedData[articlePath] = {};
-          collectedData[articlePath][percentage] = count;
-        }
-      }
-    });
-
-    var sortable = [];
-    Object.keys(collectedData).forEach((path) => {
-      let read25 = 0;
-      let readFullArticleCount = 0;
-      Object.keys(collectedData[path]).forEach((pct) => {
-        if (pct === '25%') {
-          read25 = collectedData[path][pct];
-        }
-        if (pct === '100%') {
-          readFullArticleCount = collectedData[path][pct];
-        }
-      });
-      let readFullArticlePct = 0;
-      if (read25 > 0) {
-        readFullArticlePct = (readFullArticleCount / read25) * 100;
-      }
-      collectedData[path]['readFull'] = Math.round(readFullArticlePct);
-      sortable.push([path, readFullArticlePct]);
-    });
-
-    sortable.sort(function (a, b) {
-      return b[1] - a[1];
-    });
-
-    setSortedReadingDepth(sortable);
-    sortedReadingDepth.map((item) => {
-      readingDepthRows.push(
-        <tr>
-          <td>{item[0]}</td>
-          <td>{readingDepthData[item[0]]['25%']}</td>
-          <td>{readingDepthData[item[0]]['50%']}</td>
-          <td>{readingDepthData[item[0]]['75%']}</td>
-          <td>{readingDepthData[item[0]]['100%']}</td>
-          <td>{readingDepthData[item[0]]['readFull']}%</td>
-        </tr>
-      );
-    });
-    console.log('readingDepthRows:', readingDepthRows);
-    setReadingDepthTableRows(readingDepthRows);
-    setReadingDepthData(collectedData);
-  };
-
   useEffect(() => {
     const sessionsMetric = 'ga:sessions';
     const usersMetric = 'ga:users';
@@ -271,14 +206,6 @@ const Report = (props) => {
       'ga:eventLabel',
       'ga:pagePath',
     ];
-
-    getMetricsData(viewID, startDate, endDate, eventMetrics, eventDimensions, {
-      filters: 'ga:eventCategory==NTG Article Milestone',
-    })
-      .then((resp) => {
-        displayReadingDepthResults(resp, readingDepthData, setReadingDepthData);
-      })
-      .catch((error) => console.error(error));
   }, []);
 
   return (
@@ -401,30 +328,11 @@ const Report = (props) => {
           </tbody>
         </table>
       </section>
-
-      <section className="section">
-        <div className="content">
-          <p className="subtitle is-5">Reading Depth</p>
-
-          <table className="table is-fullwidth" style={{ width: '100%' }}>
-            <thead>
-              <tr>
-                <th></th>
-                <th colSpan="4">Percentage Read</th>
-              </tr>
-              <tr>
-                <th>Article</th>
-                <th>25%</th>
-                <th>50%</th>
-                <th>75%</th>
-                <th>100%</th>
-                <th>Read Full</th>
-              </tr>
-            </thead>
-            <tbody>{readingDepthTableRows}</tbody>
-          </table>
-        </div>
-      </section>
+      <ReadingDepthData
+        viewID={viewID}
+        startDate={startDate}
+        endDate={endDate}
+      />
 
       <section className="section"></section>
       <ReadingFrequencyData
